- bracket picks saving and updating next round and refreshing are working!

+++ database/migrations/2016_05_19_201239_user_bracket_picks.php
@@ -18,8 +18,8 @@ class UserBracketPicks extends Migration
 			$table->integer('user_id')->references('id')->on('users');
 			$table->integer('bracket_game_id')->references('id')->on('bracket_games');
 			$table->integer('pool_entry_1_id')->nullable();
-			$table->double('pool_entry_2_id')->nullable();
-			$table->double('pool_entry_winner_id')->nullable();
+			$table->integer('pool_entry_2_id')->nullable();
+			$table->integer('pool_entry_winner_id')->nullable();
 			$table->timestamps();
 		});
 	}
